+++ tests/io/test_models.py
@@ -320,3 +320,26 @@ class TestAnnotationCollectionModel:
         # not exporting parent leads to it looking like the non-parent version
         new_model_without_parent = AnnotationCollectionModel.from_annotation_collection(ac, export_parent=False)
         assert new_model_without_parent == model_without_parent
+
+    def test_dump_annotation_collection(self):
+        """Model dumping should retain parent information, if it exists, and should work
+        on AnnotationCollection as well as AnnotationCollectionModel"""
+        dumped = AnnotationCollectionModel.Schema().dump(model_with_parent.to_annotation_collection())
+        assert dumped["parent_or_seq_chunk_parent"] is not None
+        dumped = AnnotationCollectionModel.Schema().dump(model_without_parent.to_annotation_collection())
+        assert dumped["parent_or_seq_chunk_parent"] is None
+        dumped = AnnotationCollectionModel.Schema().dump([model_with_parent.to_annotation_collection()], many=True)
+        assert dumped[0]["parent_or_seq_chunk_parent"] is not None
+        dumped = AnnotationCollectionModel.Schema().dump([model_without_parent.to_annotation_collection()], many=True)
+        assert dumped[0]["parent_or_seq_chunk_parent"] is None
+
+    def test_dump_annotation_collection_model(self):
+        """Model dumping should retain parent information, if it exist"""
+        dumped = AnnotationCollectionModel.Schema().dump(model_with_parent)
+        assert dumped["parent_or_seq_chunk_parent"] is not None
+        dumped = AnnotationCollectionModel.Schema().dump(model_without_parent)
+        assert dumped["parent_or_seq_chunk_parent"] is None
+        dumped = AnnotationCollectionModel.Schema().dump([model_with_parent], many=True)
+        assert dumped[0]["parent_or_seq_chunk_parent"] is not None
+        dumped = AnnotationCollectionModel.Schema().dump([model_without_parent], many=True)
+        assert dumped[0]["parent_or_seq_chunk_parent"] is None